requirements.txt
PyYAML==6.0.2
Jinja2==3.1.4
Markdown==3.6
